Automatically insert semicolons
When a newline is detected we check if the previous
token is a literal, identifier, true/false, nil,
this, return, or ). If it is then we insert a ;.

This has quite a few limitations/bugs but is
good enough for a first pass as long as you don't
get crazy with splitting statements and
expressions into multiple lines.